src/main/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.1
openai==0.26.1
pycparser==2.21
pydub==0.25.1
requests==2.28.2
soundfile==0.11.0
SpeechRecognition==3.9.0
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
python-dotenv==0.21.1
portaudio
pyaudio==0.2.11
